# daemon/docs/swagger.yaml
definitions:
...
  models.NodeView:
    properties:
      id:
        type: integer
      name:
        type: string
      privateHost:
        type: string
      privatePort:
        type: integer
      publicHost:
        type: string
      publicPort:
        type: integer
      sftpPort:
        type: integer
    type: object
  models.NodesView:
    items:
      properties:
        id:
          type: integer
        name:
          type: string
        privateHost:
          type: string
        privatePort:
          type: integer
        publicHost:
          type: string
        publicPort:
          type: integer
        sftpPort:
          type: integer
      type: object
    type: array
...
  response.Empty:
    type: object
  response.Error:
    properties:
      error:
        $ref: '#/definitions/shared.Error'
        type: object
    type: object
  shared.Error:
    properties:
      code:
        type: string
      metadata:
        type: object
      msg:
        type: string
    type: object
...
info:
  contact:
    name: PufferPanel
    url: https://pufferpanel.com
  description: PufferPanel daemon service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Pufferd API
  version: "2.0"
paths:
...
  /nodes:
    get:
      consumes:
      - application/json
      description: Gets all nodes registered to the panel
      produces:
      - application/json
      responses:
        "200":
          description: Nodes
          schema:
            $ref: '#/definitions/models.NodesView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get nodes
    post:
      consumes:
      - application/json
      description: Creates a node
      parameters:
      - description: Node Identifier
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node created
          schema:
            $ref: '#/definitions/models.NodeView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Create node
  /nodes/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes the node
      parameters:
      - description: Node Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/response.Empty'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Deletes a node
    get:
      consumes:
      - application/json
      description: Gets information about a single node
      produces:
      - application/json
      responses:
        "200":
          description: Nods
          schema:
            $ref: '#/definitions/models.NodeView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get node
    put:
      consumes:
      - application/json
      description: Updates a node with given information
      parameters:
      - description: Node Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/response.Empty'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update node
...
swagger: "2.0"